feat: Expose public top-level API
This PR creates an index.js to establish a top level, public API for the
library. It also updates the example to use the npm library.

Open to exporting other stuff in the public API. I'm not yet sure what is
meant to be exposed and what isn't.

> **Note**: I bumped the version to 0.2.0 and you'll need to publish it
for the README example to work.